requirements.txt
pydantic==2.10.6
pyperclip==1.9.0
python-dotenv==1.0.1
questionary==2.1.0
rich==13.9.4
keyring==24.3.0
openai==1.72.0
requests==2.31.0